+++ tests/unittests/test_pick_action.py
@@ -31,17 +31,24 @@ class PickActionTest(unittest.TestCase):
         task3 = todotxt.Task("(A) Todo 3")
         self.assertEqual([task3, task2, task1], pick_action.next_actions([task1, task2, task3]))
 
-    def test_completed_task_is_ignored(self):
+    def test_ignore_completed_task(self):
         """ If there's one completed and one uncompleted task, the uncompleted one is the next action. """
         completed_task = todotxt.Task("x Completed")
         uncompleted_task = todotxt.Task("Todo")
         self.assertEqual([uncompleted_task], pick_action.next_actions([completed_task, uncompleted_task]))
 
-    def test_completed_tasks_only(self):
-        """ If all tasks are completed, there's no next action. """
+    def test_ignore_future_task(self):
+        """ Ignore tasks with a start date in the future. """
+        future_task = todotxt.Task("(A) 9999-01-01 Start preparing for five-digit years")
+        regular_task = todotxt.Task("(B) Look busy")
+        self.assertEqual([regular_task], pick_action.next_actions([future_task, regular_task]))
+
+    def test_ignore_these_tasks(self):
+        """ If all tasks are completed or future tasks, there's no next action. """
         completed_task1 = todotxt.Task("x Completed")
         completed_task2 = todotxt.Task("x Completed too")
-        self.assertEqual([], pick_action.next_actions([completed_task1, completed_task2]))
+        future_task = todotxt.Task("(A) 9999-01-01 Start preparing for five-digit years")
+        self.assertEqual([], pick_action.next_actions([completed_task1, completed_task2, future_task]))
 
     def test_context(self):
         """ Test that the next action can be limited to a specific context. """